сделал отдельную авторизацию по токену спутника

+++ db/archive/20241121153747_create_private_chats_table.sql
@@ -2,14 +2,13 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS "private_chats"(
     "id"  serial NOT NULL PRIMARY KEY,
-    "user1_id" NUMERIC NOT NULL,
-    "user2_id" NUMERIC NOT NULL,
-    "msg_id" NUMERIC NOT NULL,
+    "user1_id" INTEGER NOT NULL,
+    "user2_id" INTEGER NOT NULL,
+    "msg_id" INTEGER NOT NULL,
     "created_at" TIMESTAMP NOT NULL,
     "updated_at" TIMESTAMP NOT NULL
 );
 
--- ALTER TABLE myusers ADD CONSTRAINT unique_login UNIQUE (login);
 -- +goose StatementEnd
 
 -- +goose Down
